ci: trigger CI run
